GH-#8 update `check` API of semantic analysis
**NOTE**: this doesn't *fix* our semantic analysis, just takes the first
shareable step: renames the top-level API signature to clarify what the semantic
analysis *should* be iterating over (ie: what our `ast`/LRM thinks an EqEq
program *is*).

**Next step**: walk through each `contexts` and `finds` reference in this file
and update it to match our real AST, and actually perform checks that we really
want to perform (eg: "is this `find` applying to an actual context we know
about?).